# docs/openapi.yaml
components:
  schemas:
...
    CreateScheduleRequest:
      example:
        cron_expr: 0 2 * * *
        name: nightly-fleet
        tags:
          - prod
          - critical
        template:
          input: list:fleet.txt
          plugins:
            - modbus
            - s7
            - bacnet
        timezone: Europe/Madrid
      properties:
        audit_retention_days:
          description: v1.89+ per-schedule retention override. 0=inherit global, >0=keep N days.
          maximum: 3650
          minimum: 0
          type: integer
        cron_expr:
          description: 'Cadence: cron-based (5-field). Mutually exclusive with interval_seconds.'
          example: 0 2 * * *
          type: string
        interval_seconds:
          description: 'Cadence: interval-based. Mutually exclusive with cron_expr.'
          maximum: 604800
          minimum: 60
          type: integer
        name:
          example: nightly-fleet
          type: string
        tags:
          description: v2.4+ grouping labels.
          items:
            pattern: ^[a-z0-9_-]{1,32}$
            type: string
          maxItems: 10
          type: array
        template:
          $ref: '#/components/schemas/SubmitRequest'
        timezone:
          description: IANA zone for cron evaluation. Ignored for interval schedules.
          example: Europe/Madrid
          type: string
      required:
        - name
        - template
      type: object
    Envelope:
      properties:
        data: {}
        schema:
          example: api:v1
          type: string
      required:
        - schema
        - data
      type: object
...
    ScanSchedule:
      example:
        audit_retention_days: 30
        created_at: "2026-05-15T08:00:00Z"
        enabled: true
        id: abc123def4567890
        interval_seconds: 3600
        name: nightly-fleet
        next_fire_at: "2026-05-18T09:00:00Z"
        operator: alice@example.com
        tags:
          - critical
          - prod
        template:
          input: list:fleet.txt
          plugins:
            - modbus
            - s7
            - bacnet
        timezone: ""
        updated_at: "2026-05-15T08:00:00Z"
      properties:
        audit_retention_days:
          description: v1.89+ per-schedule retention override; 0 = inherit global.
          example: 30
          type: integer
        created_at:
          format: date-time
          type: string
        cron_expr:
          example: 0 2 * * *
          type: string
        enabled:
          example: true
          type: boolean
        id:
          example: abc123def4567890
          type: string
        interval_seconds:
          example: 3600
          type: integer
        last_fired_at:
          format: date-time
          type: string
        name:
          example: nightly-fleet
          type: string
        next_fire_at:
          description: v1.77+ computed; not persisted.
          format: date-time
          type: string
        operator:
          example: alice@example.com
          type: string
        tags:
          description: 'v2.4+ grouping labels (canonicalised: sorted + deduped).'
          example:
            - prod
            - critical
          items:
            type: string
          type: array
        template:
          $ref: '#/components/schemas/SubmitRequest'
        timezone:
          example: Europe/Madrid
          type: string
        updated_at:
          format: date-time
          type: string
      required:
        - id
        - name
        - template
        - enabled
        - created_at
        - updated_at
      type: object
...
    SubmitRequest:
      properties:
        default_port:
          type: integer
        input:
          example: list:fleet.txt
          type: string
        plugins:
          items:
            type: string
          type: array
      required:
        - input
      type: object
...
info:
  description: Read-only HTTP API for ElSereno's dashboard. All endpoints return JSON with a `schema` discriminator so downstream tools can version against a concrete contract. Bearer token auth (ADR-014) gates /api/v1/* when the dashboard's auth mode is enabled.
  title: ElSereno HTTP API
  version: dev
openapi: 3.1.0
paths:
  /api/v1/findings:
    get:
      description: 'Query params: severity, protocol, min_score, created_after (RFC3339), limit (clamped 1..500, default 50). Returned order is created_at DESC; paginate by passing the oldest returned row''s created_at back as created_after.'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: Envelope payload.
        "503":
          description: DB backend unavailable.
      summary: List findings (DB-backed, cursor-paginated).
      tags:
        - findings
...
  /api/v1/runs:
    get:
      description: 'Query params: status, started_after (RFC3339), limit (1..100, default 20).'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: Envelope payload.
        "503":
          description: DB backend unavailable.
      summary: List recent scan runs with per-run finding counts.
      tags:
        - runs
  /api/v1/schedules:
    get:
      description: 'Returns every schedule sorted by name. `next_fire_at` is computed per response (v1.77+). v2.4+: optional `?tag=<value>` filters via array-overlap on the tags column.'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: Envelope payload.
      summary: List scan schedules (v1.70+, v2.4 ?tag= filter).
      tags:
        - schedules
    post:
      description: Cadence is mutually-exclusive interval_seconds OR cron_expr (v1.73+). Optional timezone (IANA, v1.75+). Optional audit_retention_days (v1.89+) per-schedule retention override (0=inherit global).
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanSchedule'
          description: Schedule created.
        "400":
          description: Validation error (name/template/cadence/cron/tz/retention).
        "503":
          description: Schedule store unavailable.
      summary: Create a scan schedule.
      tags:
        - schedules
...
  /api/v1/stream:
    get:
      description: 'text/event-stream with `event:`, `id:`, `data:` lines. Event kinds: `finding`, `run_start`, `run_end`, `audit`. Payloads are JSON objects; schema per-kind is documented in internal/web/stream/findings_bridge.go.'
      responses:
        "200":
          description: SSE-framed event stream.
        "503":
          description: Live feed unavailable (broadcaster not wired).
      summary: Server-Sent Events feed (findings / runs / audit).
      tags:
        - stream
  /api/v1/triage:
    get:
      description: Returns an array of {severity, count} objects, ordered critical → info. Empty severities are omitted.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
          description: Envelope payload.
        "503":
          description: DB backend unavailable.
      summary: Per-severity finding counts for the triage panel.
      tags:
        - triage
...
servers:
  - url: http://127.0.0.1:8787