# internal/web/server.go
package web

import (
	"context"
	"crypto/rand"
	"encoding/base64"
	"errors"
	"fmt"
	"net/http"
	"time"

	"local/elsereno/internal/config"
	"local/elsereno/internal/creds"
	"local/elsereno/internal/repo"
	"local/elsereno/internal/scanorch"
	"local/elsereno/internal/web/auth"
	"local/elsereno/internal/web/handlers"
	"local/elsereno/internal/web/httpctx"
	"local/elsereno/internal/web/stream"
)

// Options configures the HTTP server. Nil fields fall back to defaults.
type Options struct {
	Addr string
	Web  config.WebConfig
	// Vault is used to derive the CSRF key via HKDF (ADR-017). It must
	// be unlocked before Run.
	Vault *creds.Vault

	TLSCert string
	TLSKey  string

	// NowSource is the clock used for /readyz timestamps. Nil -> time.Now.
	NowSource func() time.Time

	// Querier (optional, v1.2+) backs the DB-read endpoints
	// `/api/v1/findings`, `/runs`, `/triage`. If nil the
	// endpoints return 503 and the dashboard renders the
	// skeleton-only panels.
	Querier repo.Querier

	// ScanStore (optional, v1.61+) backs the scan-orchestration
	// endpoints under `/api/v1/scans/`. If nil the endpoints
	// return 503. Production wiring passes a NewDBStore(pool)
	// (jobs survive restart); tests + dev pass a
	// NewMemoryStore() (jobs lost on restart).
	ScanStore scanorch.Store

	// ScheduleStore (optional, v1.70+) backs the scan-schedule
	// endpoints under `/api/v1/schedules/`. Nil → 503.
	ScheduleStore scanorch.ScheduleStore

	// ScheduleAuditStore (optional, v1.84+) backs the
	// force-overwrite audit log. Nil = audit-disabled
	// (force-overwrite PUTs still succeed but no row is
	// persisted, and GET /schedules/{id}/audit returns 503).
	ScheduleAuditStore scanorch.ScheduleAuditStore

	// Broadcaster (optional, v1.63+) overrides the per-server
	// SSE broadcaster. cmd_serve builds it up-front so the
	// scan-orchestration wrapper can publish events on the same
	// bus the dashboard listens to. nil means NewServer creates
	// its own internal broadcaster (back-compat with v1.62-).
	Broadcaster *stream.Broadcaster

	// PoolStatter (optional, v2.58+) is the *pgxpool.Pool that
	// backs GET /api/v1/health/pool. Nil → endpoint returns
	// 503 (memory-mode + pre-DB-bootstrap deployments).
	// cmd_serve adapts the pgxpool.Pool.Stat() snapshot into
	// the handlers.PoolStat shape via a tiny shim so this
	// package stays free of the pgxpool import.
	PoolStatter handlers.PoolStatter

	// AuthVerifier (optional, v2.59+) enforces OIDC bearer-
	// token validation on /api/v1/* routes. Nil OR
	// Enabled()==false → back-compat dev mode (X-Operator
	// header carries identity, no token check). cmd_serve
	// constructs from cfg.Auth.OIDC.* when fully configured.
	AuthVerifier *auth.Verifier

	// MetricsHandler (optional, v2.60+) is the
	// `promhttp.Handler` mounted at `GET /metrics`.
	// Constructed in cmd_serve from telemetry.Global().Handler()
	// after registering any deployment-specific collectors
	// (PoolCollector etc.). Nil → no /metrics route is
	// installed (operators can run without Prometheus).
	MetricsHandler http.Handler
}

// Server is the wrapped http.Server with the full set of timeouts and
// the default handler tree mounted.
type Server struct {
	opts      Options
	server    *http.Server
	startedAt time.Time

	csrfKey []byte

	// broadcaster is the process-local SSE fan-out. Every publisher
	// (scanner, audit writer, offensive verbs) calls into this to
	// push an event to connected `/api/v1/stream` clients.
	broadcaster *stream.Broadcaster
}

// NewServer constructs a Server with ADR-014/ADR-017 defaults applied.
func NewServer(opts Options) (*Server, error) {
	if opts.Vault == nil {
		return nil, errors.New("web: Options.Vault is required")
	}
	if opts.NowSource == nil {
		opts.NowSource = time.Now
	}
	if opts.Addr == "" {
		opts.Addr = "127.0.0.1:8787"
	}

	// Derive CSRF key via HKDF from the vault master key (ADR-017).
	csrfKey := make([]byte, 32)
	if err := opts.Vault.Derive("elsereno/csrf/v1", csrfKey); err != nil {
		return nil, fmt.Errorf("web: derive csrf key: %w", err)
	}

	broadcaster := opts.Broadcaster
	if broadcaster == nil {
		broadcaster = stream.New(128)
	}
	s := &Server{
		opts:        opts,
		startedAt:   opts.NowSource(),
		csrfKey:     csrfKey,
		broadcaster: broadcaster,
	}

	mux := http.NewServeMux()
	mux.HandleFunc("/healthz", s.healthz)
	mux.HandleFunc("/readyz", s.readyz)
	// v2.60: /metrics exposes the Prometheus expo. Optional,
	// when MetricsHandler is nil the route stays unbound (404
	// in default deployments that don't run Prometheus).
	if opts.MetricsHandler != nil {
		mux.Handle("/metrics", opts.MetricsHandler)
	}
	mux.Handle("/api/v1/", handlers.APIV1(handlers.APIV1Deps{
		Broadcaster:        s.broadcaster,
		Querier:            opts.Querier,
		ScanStore:          opts.ScanStore,
		ScheduleStore:      opts.ScheduleStore,
		ScheduleAuditStore: opts.ScheduleAuditStore,
		// v2.58: forward the pgxpool stat surface.
		PoolStatter: opts.PoolStatter,
		// v2.59: forward the OIDC verifier (nil = back-compat).
		AuthVerifier: opts.AuthVerifier,
	}))
	mux.Handle("/admin/security", handlers.Security())
	mux.Handle("/", handlers.Dashboard())

	// #nosec G112 -- all timeouts are set explicitly below.
	s.server = &http.Server{
		Addr:              opts.Addr,
		Handler:           securityHeaders(mux),
		ReadHeaderTimeout: nonzero(opts.Web.ReadHeaderTimeout, 5*time.Second),
		ReadTimeout:       nonzero(opts.Web.ReadTimeout, 30*time.Second),
		WriteTimeout:      nonzero(opts.Web.WriteTimeout, 30*time.Second),
		IdleTimeout:       nonzero(opts.Web.IdleTimeout, 120*time.Second),
		MaxHeaderBytes:    1 << 14, // 16 KiB
	}
	return s, nil
}

// Run starts the server and blocks until ctx is cancelled or the
// server errors.
func (s *Server) Run(ctx context.Context) error {
	errCh := make(chan error, 1)
	go func() {
		if s.opts.TLSCert != "" && s.opts.TLSKey != "" {
			errCh <- s.server.ListenAndServeTLS(s.opts.TLSCert, s.opts.TLSKey)
			return
		}
		errCh <- s.server.ListenAndServe()
	}()

	select {
	case <-ctx.Done():
		// Decoupled timeout: ctx is already cancelled, so we cannot
		// propagate it to Shutdown or the drain would abort
		// immediately. We want the drain window to finish.
		shutdownCtx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
		defer cancel()
		//nolint:contextcheck // decoupled on purpose: parent ctx is already cancelled
		_ = s.server.Shutdown(shutdownCtx)
		return ctx.Err()
	case err := <-errCh:
		return err
	}
}

// Addr returns the bound address (for tests; Run must have started).
func (s *Server) Addr() string { return s.server.Addr }

// Handler exposes the assembled http.Handler so tests can
// fire requests via httptest without binding a real port.
// Stable since v2.60+.
func (s *Server) Handler() http.Handler { return s.server.Handler }

// Broadcaster returns the SSE broadcaster so the process's publishers
// (scanner, audit writer, offensive verbs) can push events to
// connected /api/v1/stream clients.
func (s *Server) Broadcaster() *stream.Broadcaster { return s.broadcaster }

func (s *Server) healthz(w http.ResponseWriter, _ *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	_, _ = fmt.Fprintf(w, `{"status":"ok","started_at":%q}`, s.startedAt.UTC().Format(time.RFC3339))
}

func (s *Server) readyz(w http.ResponseWriter, r *http.Request) {
	// ADR-022 shape: /readyz returns 200 with a JSON status
	// when the server is ready to take traffic, 503 when a
	// critical dependency is down. "Critical" in v1.2 is the
	// DB pool IF configured, an unconfigured pool is NOT a
	// failure (the file-backed audit + SSE still work).
	//
	// Audit readiness is reported separately: if the Querier is
	// wired we assume the caller's `serve` wire-up also opened
	// the audit writer; verifying the writer directly would
	// require plumbing it through to the Server, which is a
	// v1.3 scope expansion.
	ctx, cancel := context.WithTimeout(r.Context(), 500*time.Millisecond)
	defer cancel()

	db := "skipped"
	overall := "ready"
	httpStatus := http.StatusOK
	if s.opts.Querier != nil {
		var one int
		err := s.opts.Querier.QueryRow(ctx, `SELECT 1`).Scan(&one)
		switch {
		case err == nil && one == 1:
			db = "ok"
		default:
			db = "down"
			overall = "degraded"
			httpStatus = http.StatusServiceUnavailable
		}
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(httpStatus)
	_, _ = fmt.Fprintf(w, `{"status":%q,"db":%q,"audit":"skipped","uptime_s":%d}`,
		overall, db, int(time.Since(s.startedAt).Seconds()))
}

// securityHeaders adds the default header set (HSTS only under TLS so
// it is the caller's responsibility to wrap with https middleware).
// A CSP nonce is generated per request and stashed in the request
// context (see `internal/web/httpctx`) so templates can render
// inline <script nonce=...> without weakening the policy.
func securityHeaders(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		h := w.Header()
		nonce := randNonce()
		h.Set("Content-Security-Policy", "default-src 'self'; script-src 'self' 'nonce-"+nonce+"'; style-src 'self' 'nonce-"+nonce+"'; base-uri 'none'; frame-ancestors 'none'; connect-src 'self'")
		h.Set("X-Frame-Options", "DENY")
		h.Set("X-Content-Type-Options", "nosniff")
		h.Set("Referrer-Policy", "no-referrer")
		h.Set("Permissions-Policy", "camera=(), microphone=(), geolocation=()")
		if r.TLS != nil {
			h.Set("Strict-Transport-Security", "max-age=63072000; includeSubDomains")
		}
		r = r.WithContext(httpctx.WithCSPNonce(r.Context(), nonce))
		next.ServeHTTP(w, r)
	})
}

// randNonce returns a short base64 value suitable for a CSP nonce.
func randNonce() string {
	var b [12]byte
	_, _ = rand.Read(b[:])
	return base64.RawURLEncoding.EncodeToString(b[:])
}

func nonzero[T comparable](a, def T) T {
	var zero T
	if a == zero {
		return def
	}
	return a
}